requirements.txt
feedparser>=6.0
beautifulsoup4>=4.12
httpx>=0.27
aiohttp>=3.9
python-dateutil>=2.9
sentence-transformers>=2.7
einops>=0.8
sqlalchemy>=2.0
trafilatura>=2.0
spacy>=3.7
